Use uint8_t for TokenKind, re-enable enum size check
Current value set fits in 1 byte. Will change to uint16_t when
the enum actually exceeds 255 values.

+++ src/lexer/token.h
@@ -8,7 +8,7 @@
 
 namespace delta {
 
-enum class TokenKind : uint16_t {
+enum class TokenKind : uint8_t {
   // Sentinel
   Eof = 0,
   Error,
